Add files via upload
The algorithm that I have attached is made to sort the numbers that are randomly in between the smallest and greatest number. However, it sorts it from the smallest to greatest, in the first radix pass, it sorts it by the '1's position, then in the second radix pass, it sorts it by the '10's position, then the 3rd, '100's, then 4th, s,1,000's, and it will just keep going until there are no more positions in the greatest number given and prompted by the user.