Add files via upload
I decided to add a couple files with a couple other coding projects as well. These files were made when I was making card games with friends. What we would do is try and write code that would be like card games that we see in apps such as solitaire, blackjack, or even Uno. In this particular code, I didn't quite get it to work, but I was the one responsible for deciding who's turn it was while playing in the game. It was a fun project with friends, as well as a great way to try and learn more and improve in our coding abilities.